app/backend/requirements.txt
google-adk>=0.0.5
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.23.2
# react>=18.2.0 # This is a frontend dependency
# tailwindcss>=3.3.5 # This is a frontend dependency
pymongo>=4.5.0
pydantic>=2.4.2
websockets>=12.0.0
python-socketio>=5.10.0
litellm>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.5.0
# UPDATED: Dependencies for file processing
pandas>=2.0.0
openpyxl>=3.1.2
pypdf2>=3.0.0
python-multipart>=0.0.6 